scorpion/requirements.txt
exifread==3.0.0
Pillow==9.2.0
